+++ tests/dtypes/test_dtype.cpp
@@ -21,7 +21,7 @@ TEST(DType, BasicConstructor) {
     // Default-constructed dtype properties
     EXPECT_EQ(pattern_type_id, d.type_id());
     EXPECT_EQ(pattern_kind, d.kind());
-    EXPECT_EQ(1, d.alignment());
+    EXPECT_EQ(1u, d.alignment());
     EXPECT_EQ(0u, d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -29,7 +29,7 @@ TEST(DType, BasicConstructor) {
     d = dtype(bool_type_id);
     EXPECT_EQ(bool_type_id, d.type_id());
     EXPECT_EQ(bool_kind, d.kind());
-    EXPECT_EQ(1, d.alignment());
+    EXPECT_EQ(1u, d.alignment());
     EXPECT_EQ(1u, d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -37,14 +37,14 @@ TEST(DType, BasicConstructor) {
     d = dtype(int8_type_id);
     EXPECT_EQ(int8_type_id, d.type_id());
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ(1, d.alignment());
+    EXPECT_EQ(1u, d.alignment());
     EXPECT_EQ(1u, d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
     // int16 dtype
     d = dtype(int16_type_id);
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ(2, d.alignment());
+    EXPECT_EQ(2u, d.alignment());
     EXPECT_EQ(2u, d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -52,7 +52,7 @@ TEST(DType, BasicConstructor) {
     d = dtype(int32_type_id);
     EXPECT_EQ(int32_type_id, d.type_id());
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ(4, d.alignment());
+    EXPECT_EQ(4u, d.alignment());
     EXPECT_EQ(4u, d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -60,14 +60,14 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<int>();
     EXPECT_EQ(int32_type_id, d.type_id());
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ((int)sizeof(int), d.alignment());
+    EXPECT_EQ(sizeof(int), d.alignment());
     EXPECT_EQ(sizeof(int), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
     // long
     d = make_dtype<long>();
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ((int)sizeof(long), d.alignment());
+    EXPECT_EQ(sizeof(long), d.alignment());
     EXPECT_EQ(sizeof(long), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -75,7 +75,7 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<long long>();
     EXPECT_EQ(int64_type_id, d.type_id());
     EXPECT_EQ(int_kind, d.kind());
-    EXPECT_EQ((int)sizeof(long long), d.alignment());
+    EXPECT_EQ(sizeof(long long), d.alignment());
     EXPECT_EQ(sizeof(long long), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -83,14 +83,14 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<unsigned int>();
     EXPECT_EQ(uint32_type_id, d.type_id());
     EXPECT_EQ(uint_kind, d.kind());
-    EXPECT_EQ((int)sizeof(unsigned int), d.alignment());
+    EXPECT_EQ(sizeof(unsigned int), d.alignment());
     EXPECT_EQ(sizeof(unsigned int), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
     // unsigned long
     d = make_dtype<unsigned long>();
     EXPECT_EQ(uint_kind, d.kind());
-    EXPECT_EQ((int)sizeof(unsigned long), d.alignment());
+    EXPECT_EQ(sizeof(unsigned long), d.alignment());
     EXPECT_EQ(sizeof(unsigned long), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -98,7 +98,7 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<unsigned long long>();
     EXPECT_EQ(uint64_type_id, d.type_id());
     EXPECT_EQ(uint_kind, d.kind());
-    EXPECT_EQ((int)sizeof(unsigned long long), d.alignment());
+    EXPECT_EQ(sizeof(unsigned long long), d.alignment());
     EXPECT_EQ(sizeof(unsigned long long), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -106,7 +106,7 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<float>();
     EXPECT_EQ(float32_type_id, d.type_id());
     EXPECT_EQ(real_kind, d.kind());
-    EXPECT_EQ((int)sizeof(float), d.alignment());
+    EXPECT_EQ(sizeof(float), d.alignment());
     EXPECT_EQ(sizeof(float), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 
@@ -114,7 +114,7 @@ TEST(DType, BasicConstructor) {
     d = make_dtype<double>();
     EXPECT_EQ(float64_type_id, d.type_id());
     EXPECT_EQ(real_kind, d.kind());
-    EXPECT_EQ((int)sizeof(double), d.alignment());
+    EXPECT_EQ(sizeof(double), d.alignment());
     EXPECT_EQ(sizeof(double), d.element_size());
     EXPECT_EQ(NULL, d.extended());
 }
